chore: code cleanup of guidebook graph logic
No new or altered logic here. This PR only moves code around, reducing the graph.ts monolith.
